Began implementation of general output parser for ADF.
There has been a lot of issues arising from ADF changing
their printing of stdout. This has come from changing to
the new AMS implementation. This will serve as the
groundwork to identify which version of ADF we are using
and at the same time be able to handle the parsing logic
on the multiple modules that they have, such as NMR.