Bug fixes and clarification
Fixed a bug with the ADF excitations parser.
Copied the parser from the AMS section as
that has not changed from the ADF 2017 version.

Clarified the tensors that are parsed from the
CPL section. Added a keyword to change between
the two that are parsed. Defaults to the Cartesian
to align with what is on the Tape21 file.

Added the ability to pass keyword arguments to the
general output parser due to this change.

Added tests for this new behavior.