Added parsers for tddft results from gaussian
Added parsers for the electric dipole and quadrupole
moment, and magnetic dipole moment. The electric
dipole parser can read the data in the length or
velocity formalism with a keyword.

Also added a parser for all spectral properties
including, rotatory and dipole strengths in the length
and velocity formalisms.

Added relevant unit tests.