Should fix reporting to codacy
Changed some of the internal working so that even
if the unit tests fail the artifact with the
coverage information will be uploaded so that it
can be read to then upload it to codacy.

Issue was that if the test failed the artifact
was not written and in the next step it is
looking to download a file that does not exist.