Version lock ipywidgets add pull-request specific workflow
Version locked ipywidgets to less than 8.0 as
a method that we use has been deprecated. This
will ensure that the tests in the widget
module pass.

Added a specific pull-request workflow that
does not do anything to upload coverage to
Codacy as the secret variables that are
needed are protected by definition.
Code coverage will be triggered only when
the pull request is merged into main.